Proxy failed requests to backend for development
Since the front-end development server runs on a seperate host domain to
the back-end development server, we need to proxy requests to the
back-end path (in production, both the back-end and front-end code will
have the same host, so this is not a problem).

This is mainly to avoid problems with CORS access control during
development. See here:

https://create-react-app.dev/docs/proxying-api-requests-in-development/

for more info.