Add functions to return reviews by user or movie
Split the getReviews function in Reviews.java and ReviewDB.java into two
functions that allow getting by the user id or getting by the movie id
(getReviewByUserId or getReviewByMovieId respectively). This resolves a
bug where the dashboard endpoint was not receiving the list of reviews
associated with the user.

As a side note, The functionality of Reviews is a bit all over the
place. Reviews simultaneously contains methods for fetching from the
review database, but also represents a type (List of Reviews). I think
it would be a good idea to consider the one purpose this object
represents, to avoid duplicate functionality and people not knowing
which object to use to get stuff.