Modify dashboard page loading behaviour and layout
Load the dashboard in one go, waiting for the dashboard data to be
fetched from the server. Also, the <Header /> component now accepts a
property (props.isLoggedIn) that can be used by the pages on which it is
included to indicate whether the user is currently logged in or not.

Also add some stub locations (<IfPending/> and <IfRejected>) where
elements can be put in the case where the page is still loading or the
user was not authorized to access the data (or something went wrong).