Implement delete review function for dashboard
Link the reviews in the dashboard with the delete review API endpoint.
TODO: There is a strange bug where the review is being deleted, but the
client is still receiving 400 Bad Request error, with the body set to
'Failed to remove review.'. Need to ping Kristian to have a look at it