Refactor FilmPoker code to centralize WS comms
This refactors the FilmPoker logic to consolidate communication to the
WebSocket in the main component class (PlayFilmPoker). Along with a
clean-up of indentation, and renaming some components, we will now pass
data fetched from the websocket down to the screens/components that
require it, and also pass update callbacks so that all the state
relevant to what needs to be communicated to the socket is consolidated.