Hook up fetching of wishlist and watchlist
Fetch and render the movies from wishlist and watchlist.

TODO: There is a bug somewhere here such that if you add a movie in the
wishlist tab, then remove that movie in the watchlist tab, then add it
again in the wishlist tab, the movie won't show up in the confirmed
movies screen.