Save tab position after reload dashboard data
This commit moves the state about the tab position of the dashboard into
the parent component (Account). This is because, when reload dashboard
data is called, the Account component unmounts the Dashboard component
and replaces it with the components for pending data. This unmount
causes the state of the Dashboard data to be lost, and a new Dashboard
is mounted later with new state.

Move the state about tab position into Account so we can propogate it to
Dashboard without losing it when Dashboard unmounts.